+++ src/config.rs
@@ -12,8 +12,8 @@ use crate::{download::DownloadParams, compare::CompareParams};
 #[derive(Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
 pub struct Config {
     // Download
-    pub flex_query_id: String,
-    pub ib_token: String,
+    pub flex_query_id: Option<String>,
+    pub ib_token: Option<String>,
     // Comparison
     pub flex_report_path: Option<String>,
     pub flex_reports_dir: Option<String>,
@@ -34,15 +34,15 @@ pub fn get_dl_config(params: DownloadParams) -> Config {
     // overwrite the file values if provided by other means
     
     if let Some(query_id) = params.query_id {
-        cfg.flex_query_id = query_id.to_string();
+        cfg.flex_query_id = Some(query_id.to_string());
     } else if let Ok(env_queryid) = std::env::var("IBFLEX_QUERYID") {
-        cfg.flex_query_id = env_queryid;
+        cfg.flex_query_id = Some(env_queryid);
     };
 
     if let Some(token) = params.token {
-        cfg.ib_token = token;
+        cfg.ib_token = Some(token);
     } else if let Ok(env_tkn) = std::env::var("IBFLEX_TOKEN") {
-        cfg.ib_token = env_tkn;
+        cfg.ib_token = Some(env_tkn);
     };
 
     cfg